+++ src/JsonClassCodec.php
@@ -69,7 +69,7 @@ interface JsonClassCodec {
 	 *
 	 * @param class-string<T> $className The class we're looking for a hint for
 	 * @param string $keyName The name of the array key we'd like a hint on
-	 * @return class-string<JsonCodecable>|'stdClass'|null
+	 * @return ?class-string
 	 */
 	public function jsonClassHintFor( string $className, string $keyName ): ?string;
 }