+++ HISTORY.md
@@ -1,6 +1,8 @@
 # Release History
 
 ## x.x.x (not yet released)
+* JsonCodec::addCodecFor() is added to provide the ability to
+  serialize/deserialize objects which don't implement JsonCodecable.
 
 ## 1.0.0